+++ crates/brynja-hash-sha2/src/error.rs
@@ -14,3 +14,5 @@ error_type!(Sha224Error, "SHA-224", "64");
 error_type!(Sha256Error, "SHA-256", "64");
 error_type!(Sha384Error, "SHA-384", "128");
 error_type!(Sha512Error, "SHA-512", "128");
+error_type!(Sha512_224Error, "SHA-512/224", "128");
+error_type!(Sha512_256Error, "SHA-512/256", "128");