+++ scripts/check-zeroization-sanitizer.sh
@@ -30,3 +30,13 @@ RUSTFLAGS="-Zsanitizer=address" cargo +nightly-2026-08-16 test \
     -p brynja-hash-sha2 \
     --test sha512 \
     --target x86_64-unknown-linux-gnu
+
+RUSTFLAGS="-Zsanitizer=address" cargo +nightly-2026-08-16 test \
+    -p brynja-hash-sha2 \
+    --test sha512_224 \
+    --target x86_64-unknown-linux-gnu
+
+RUSTFLAGS="-Zsanitizer=address" cargo +nightly-2026-08-16 test \
+    -p brynja-hash-sha2 \
+    --test sha512_256 \
+    --target x86_64-unknown-linux-gnu